+++ src/Creature.h
@@ -64,7 +64,7 @@ class Creature : public ICreature {
 
   void set_lifetime(const int new_lifetime) override;
 
-  void set_vx_vy(const int vx, const int vy) override;
+  void set_vx_vy(const float vx, const float vy) override;
 };
 
 #endif  // CREATURE_H
